requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
bcrypt==4.2.1
click==8.1.7
colorama==0.4.6
Django==5.0.7
django-rest-framework==0.1.0
djangorestframework==3.15.2
fastapi==0.115.5
h11==0.14.0
idna==3.10
neo4j==5.26.0
passlib==1.7.4
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.32.1
